things

+++ demo/main.js
@@ -12,6 +12,10 @@
             .filter(m => m.classList.contains(classname))
             .filter(m => !m.classList.add('open'))[0];
 
+    const empty = (elem) => { while (elem.firstChild) elem.removeChild(elem.firstChild); },
+          remove = (elem) => { if (elem && elem.parentNode) elem.parentNode.removeChild(elem); };
+
+
     async function svc(secure, path, method, json) {
 
         const url = `${baseUrl}/${path}`,
@@ -26,7 +30,8 @@
               text = await response.text();
 
         if (response.status === 401 && secure === true) {
-            return obtain_login(arguments);
+            return obtain_login(true)
+                .then(() => svc(secure, path, method, json));
         }
         else if (!response.ok) {
             throw text;
@@ -35,12 +40,14 @@
         return JSON.parse(text);
     }
 
-    function obtain_login(args) {
+    function obtain_login(fresh) {
         
-        wipe_token();
-        const modal = show_modal('login');
+        return new Promise(resolve => {
 
-        return new Promise((resolve, reject) => {
+            if (fresh === true) wipe_token();
+            else if (load_token()) return resolve();
+
+            const modal = show_modal('login');
 
             modal.querySelector('[type=button]').onclick = () => {
 
@@ -53,7 +60,31 @@
                         passctrl.value = '';
                         save_token(json.token);
                         modal.classList.remove('open');
-                        return args && args[0] === true ? resolve(svc.apply(null, args)) : null;                   
+                        return resolve();
+                    })
+                    .catch(err => console.warn(err));
+            }
+        });
+    }
+
+    function obtain_registration() {
+
+        return new Promise(resolve => {
+
+            const modal = show_modal('register');
+
+            modal.querySelector('[type=button]').onclick = () => {
+
+                const passctrl = modal.querySelector('[placeholder=password]'),
+                      username = modal.querySelector('[placeholder=username]').value,
+                      password = passctrl.value;
+
+                svc(false, 'user', 'POST', { username, password })
+                    .then(json => {
+                        passctrl.value = '';
+                        save_token(json.token);
+                        modal.classList.remove('open');
+                        return resolve();
                     })
                     .catch(err => console.warn(err));
             }
@@ -61,16 +92,28 @@
     }
 
     const list_accounts = () => svc(true, 'account', 'GET').then(accounts => { 
-        console.log('accounts', accounts);
+        const target = qsa('#accounts > .list')[0];
+        empty(target);
+        accounts.forEach(acc => {
+            const elem = document.createElement('div');
+            elem.innerHTML = acc.accountId;
+            target.appendChild(elem);
+        });
     });
+
+    const logout = () => {
+        wipe_token();
+        qsa('.list').forEach(l => empty(l));
+        obtain_login();
+    }
 
     window.addEventListener('load', () => {
 
-        document.querySelector('#logout').onclick = obtain_login;
+        document.querySelector('#logout').onclick = logout;
         document.querySelector('#login').onclick = list_accounts;
+        document.querySelector('#register').onclick = () => obtain_registration().then(() => list_accounts());
 
-        list_accounts();
-
+        obtain_login().then(() => list_accounts());
     });
 
 })(window);
